refactor: extract useTrmnlSync hook
Moves TRMNL e-ink dashboard state and refs out of App.jsx into
src/hooks/useTrmnlSync.js:
- 4 useState: trmnlConfig, trmnlSyncStatus, trmnlLastSynced,
  trmnlMarkupCopied
- 6 useRef: trmnlSyncTimerRef, trmnlLastPushRef, trmnlBackoffUntilRef,
  trmnlBackoffCountRef, trmnlSyncInProgressRef, performTrmnlSyncRef
- Persist trmnlConfig effect (only depends on internal state)

The auto-sync effect and performTrmnlSync remain in App.jsx — they
depend on tasks, habits, routines, weather, and other external state.
trmnlBackoffCountRef and trmnlSyncInProgressRef are included even
though not listed in the plan, as they are private TRMNL refs.
